Rename FilterIndexRule & JoinIndexRule
